+++ src/main/java/org/market/app/dto/ProductsInCart.java
@@ -3,11 +3,12 @@ package org.market.app.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 @Builder
 public class ProductsInCart {
     List<ItemDto> items;
-    long totalCost;
+    BigDecimal totalCost;
 }
